requirements.txt
feedparser
python-telegram-bot==20.7
